+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+- Added `deepEquals` to deeply check for equality standard values, observable values, and tree nodes.
+
 ## 0.30.1
 
 - Fixed an issue with the readonly middleware where sometimes it was possible to write to a protected node when the write was being done from an unprotected parent node action.